apps/api/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.37
asyncpg==0.30.0
psycopg[binary]==3.2.13
alembic==1.14.1
pydantic==2.10.5
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
pwdlib[argon2]==0.2.1
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
email-validator==2.2.0
slowapi==0.1.9
itsdangerous==2.2.0
python-multipart==0.0.18
